% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptimumDA.R
\name{OptimumDA}
\alias{OptimumDA}
\title{Discriminant Analysis}
\description{
This dialog is used to specify a  discriminant analysis model to be fit by the \link[OptimClassifier]{Optim.DA} function.
}
\examples{
{
# Examples of use are provided in the help of each option of menu.
}

}
\seealso{
\code{\link[OptimClassifier]{Optim.DA}}
}
