% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptimumLM.R
\name{OptimumLM}
\alias{OptimumLM}
\title{Linear Model}
\description{
This dialog is used to specify a  linear model to be fit by the \link[OptimClassifier]{Optim.LM} function.
}
\examples{
{
# Examples of use are provided in the help of each option of menu.
}

}
\seealso{
\code{\link[Rcmdr]{linearModel}}, \code{\link[OptimClassifier]{Optim.LM}}
}
