% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptimumLMM.R
\name{OptimumLMM}
\alias{OptimumLMM}
\title{Linear Mixed Model}
\description{
This dialog is used to specify a  linear mixed model to be fit by the \link[OptimClassifier]{Optim.LMM} function.
}
\seealso{
\code{\link[OptimClassifier]{Optim.LMM}}
}
