% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptimumNN.R
\name{OptimumNN}
\alias{OptimumNN}
\title{Artificial Neural Networks}
\description{
This dialog is used to specify a neural network model to be fit by the \link[OptimClassifier]{Optim.NN} function.
}
\examples{
{
# Examples of use are provided in the help of each option of menu.
}

}
\seealso{
\code{\link[OptimClassifier]{Optim.NN}}
}
