% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptimumGLM.R
\name{OptimumGLM}
\alias{OptimumGLM}
\title{Generalized Linear Model}
\description{
This dialog is used to specify a generalized linear model to be fit by the \link[OptimClassifier]{Optim.GLM} function.
}
\seealso{
\code{\link[Rcmdr]{generalizedLinearModel}}, \code{\link[OptimClassifier]{Optim.GLM}}
}
