% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptimumSVM.R
\name{OptimumSVM}
\alias{OptimumSVM}
\title{Support Vector Machines}
\description{
This dialog is used to specify a support vector machines model to be fit by the \link[OptimClassifier]{Optim.SVM} function.
}
\seealso{
\code{\link[OptimClassifier]{Optim.SVM}}
}
